changed getSpeed to double instead of int to match types. Also change variables to protected so that the inheriting classes could use them

+++ src/Humanoid.java
@@ -8,7 +8,7 @@ public interface Humanoid
    * Get the current speed of the zombie
    * @return current speed
    */
-  int getSpeed();
+  double getSpeed();
 
   /**
    * Get the current heading of the zombie
